requirements.txt
lightrag-hku>=1.3.7
qdrant-client>=1.7.0
tree-sitter>=0.20.0
tree-sitter-pascal @ git+https://github.com/Isopod/tree-sitter-pascal.git
openai>=1.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
click>=8.0.0
python-dotenv>=1.0.0
networkx>=3.0
graspologic>=3.0